Reorganize CSS files near to their components
