+++ assets/css/style.css
@@ -236,7 +236,7 @@ th.fit button:focus {
   padding: 0 1rem;
 }
 
-.submission-traits{
+.submission-traits {
   font-size:75%;
   width:100%;
   display:flex;
