+++ edurange_refactored/templates/student_view/components/chatbox/chatbox.component.jsx
@@ -1,7 +1,8 @@
+import "./chatbox.styles.css"
 class Chatbox extends React.Component {
     render() {
         return (
-            <a href='student_view' className='chatbox'>
+            <a href='student_view' className='edu-chatbox'>
                 <p>How can we help?</p>
             </a>
         );
