+++ edurange_refactored/templates/student_view/components/question/question.styles.css
@@ -1,0 +1,82 @@
+.edu-question {
+  display: flex;
+  flex-direction: column;
+  flex: 1 0 75%;
+  padding: 1rem;
+  margin: 1rem;
+  box-shadow: 1px 2px 5px 1px lightgray;
+}
+
+.edu-submission-traits {
+  font-size:75%;
+  width:100%;
+  display:flex;
+  flex-wrap: wrap;
+  align-content: space-between;
+  display:inline-flex;
+}
+
+.edu-submission-points {
+  background-color: lightgray;
+  padding: 2px 2px 2px 2px;
+  border-radius: 5px;
+}
+
+.edu-submission-state {
+  margin-left: auto;
+  background-color: lightgray;
+  padding: .25rem .5rem;
+  border-radius: 5px;
+}
+
+.edu-question-text {
+  font-size: 90%;
+  background-color: lightgray;
+  padding: 5px;
+  border-radius: 5px;
+}
+
+.edu-answer-area {
+  display: flex;
+  flex-direction: row;
+  font-size: 75%;
+}
+
+.edu-answer {
+  box-sizing: border-box;
+  resize: none; /* prevents corner "grabber" */
+  border-radius: 4px;
+  border: 1px solid lightgray;
+  width: 100%;
+  padding: 1rem;
+  transition-duration: 0.2s;
+  transition-timing-function: ease-in-out;
+}
+
+.edu-answer:focus, .answer:hover {
+background: lightgray;
+}
+
+.edu-submit {
+  color: white;
+  background-color: #509b79;
+  border-radius: 5px;
+  border-width: 0px;
+  padding: 15px 1em;
+  text-decoration: none;
+  margin: 0 4px;
+  cursor: pointer;
+  box-shadow: 2px 2px 2px grey;
+  align-self: flex-start;
+  transition-duration: .3s;
+}
+
+.edu-submit:hover {
+  background-color: #2e9566;
+}
+
+.edu-submit:active {
+  transition-duration: 0;
+  box-shadow: 0 0 lightgray;
+  transform: translate(1px, 2px);
+}
