+++ edurange_refactored/templates/student_view/components/reading/reading.component.jsx
@@ -1,11 +1,12 @@
+import "./reading.styles.css"
 class Reading extends React.Component {
     render() {
         // TODO get readings in the json contents file rather than reading the files from the front end
         // fetch(`./readings/${this.props.fileName}`).then((resp) => {reading = resp.text()})
         return (
-            <div className='guidebox'>
-                <h3>Guidebox goes here</h3>
-                <p>More guidebox stuff here</p>
+            <div className='edu-reading'>
+                <h3>Reading Section Goes Here</h3>
+                <p>Render reading from markdown into here.</p>
             </div>
         );
     }
